requirements.txt
buttplug-py==0.2.0
dxcam_cpp==0.2.3
numpy==2.2.6
opencv-python==4.11.0.86
pynput==1.8.1
PySide6==6.9.0
